+++ MusicKit/Frameworks/SndKit/SndAudioProcessorDelay.m
@@ -98,7 +98,7 @@
 // paramValue
 ////////////////////////////////////////////////////////////////////////////////
 
-- (float) paramValue: (const int) index
+- (float) paramValue: (const NSInteger) index
 {
   float r = 0.0f;
   switch (index) {
@@ -112,7 +112,7 @@
 // paramName
 ////////////////////////////////////////////////////////////////////////////////
 
-- (NSString*) paramName: (const int) index
+- (NSString*) paramName: (const NSInteger) index
 {
   NSString *r = nil;
   
@@ -127,7 +127,7 @@
 // setParam 
 ////////////////////////////////////////////////////////////////////////////////
 
-- (void) setParam: (const int) index toValue: (const float) v
+- (void) setParam: (const NSInteger) index toValue: (const float) v
 {
   switch (index) {
     case dlyLength:
