+++ MusicKit/Frameworks/SndKit/SndAudioProcessorDistortion.m
@@ -81,7 +81,7 @@
 // getParamName
 //////////////////////////////////////////////////////////////////////////////
 
-- (NSString*) paramName: (const int) index
+- (NSString*) paramName: (const NSInteger) index
 {
   NSString *r = nil;
   switch (index)  {
@@ -113,7 +113,7 @@
 // setParam
 //////////////////////////////////////////////////////////////////////////////
 
-- (void) setParam: (const int) index toValue: (const float) value
+- (void) setParam: (const NSInteger) index toValue: (const float) value
 {
   switch (index) {
     case distort_kBoostAmount:
