+++ MusicKit/Frameworks/SndKit/SndAudioProcessorToneGenerator.m
@@ -82,7 +82,7 @@ double freqparam2freq(float freq)
 // setParam
 //////////////////////////////////////////////////////////////////////////////
 
-- (void) setParam: (const int) index toValue: (const float) value
+- (void) setParam: (const NSInteger) index toValue: (const float) value
 {
   switch (index) {
     case toneGen_kFreq:  freq     = value;
@@ -98,7 +98,7 @@ double freqparam2freq(float freq)
 // paramValue
 //////////////////////////////////////////////////////////////////////////////
 
-- (float) paramValue: (const int) index
+- (float) paramValue: (const NSInteger) index
 {
   float r = 0.0;
   switch (index) {
@@ -114,7 +114,7 @@ double freqparam2freq(float freq)
 // paramName
 //////////////////////////////////////////////////////////////////////////////
 
-- (NSString*) paramName: (const int) index
+- (NSString*) paramName: (const NSInteger) index
 {
   NSString *r = nil;
   switch (index) {
